Factored out Worker class.
